requirements.txt
fastapi
uvicorn
websockets
openai
redis
pydantic
pydantic-settings
python-dotenv
google.generativeai
